refactor: remove circular dependencies from common lib
