+++ frontend/common/src/defi/common.ts
@@ -1,0 +1,14 @@
+import type { Balance } from '../balances';
+
+export interface ProfitLossModel {
+  readonly address: string;
+  readonly asset: string;
+  readonly value: Balance;
+}
+
+export enum LpType {
+  UNISWAP_V2 = 'UNISWAP_V2',
+  UNISWAP_V3 = 'UNISWAP_V3',
+  SUSHISWAP = 'SUSHISWAP',
+  BALANCER = 'BALANCER',
+}
