+++ frontend/common/src/defi/index.ts
@@ -1,20 +1,7 @@
-import type { Balance } from '../index';
-
 export * from './xswap';
 
 export * from './aave';
 
 export * from './balancer';
 
-export interface ProfitLossModel {
-  readonly address: string;
-  readonly asset: string;
-  readonly value: Balance;
-}
-
-export enum LpType {
-  UNISWAP_V2 = 'UNISWAP_V2',
-  UNISWAP_V3 = 'UNISWAP_V3',
-  SUSHISWAP = 'SUSHISWAP',
-  BALANCER = 'BALANCER',
-}
+export * from './common';
